refactor: split MarketOrder page and container logic
